requirements.txt
Django==1.8.5
django-cities==0.4.2
django-filter==0.11.0
djangorestframework==3.3.0
drf-nested-routers==0.10.0 -e git://github.com/alanjds/drf-nested-routers@master#egg=drf-nested-routers-master
Markdown==2.6.3
psycopg2==2.6.1
wheel==0.26.0
